backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Todo Backend API with Better Auth JWT integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "httpx>=0.25.2",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "slowapi>=0.1.8",
    "python-jose[cryptography]>=3.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
]